Issue #1226 - ensure that the correct implementation classes are used
